return long in getSize function, update results in run4

+++ src/main/java/com/bdspro/databases/Database.java
@@ -15,7 +15,7 @@ public interface Database {
 
     int runQuery(String queryString);
 
-    int getSize(String datasetTableName);
+    long getSize(String datasetTableName);
 
     int getRowCount(String datasetTableName);
